pipeline: move merge test into human tests
