+++ tests/pipeline/test_human_config.py
@@ -8,11 +8,14 @@ from pathlib import Path
 
 from lenskit.als import BiasedMFScorer, ImplicitMFScorer
 from lenskit.basic import BiasScorer, FallbackScorer, TopNRanker
+from lenskit.basic.candidates import TrainingItemsCandidateSelector
 from lenskit.config import load_config_data
+from lenskit.knn.slim import SLIMScorer
 from lenskit.pipeline import Pipeline, PipelineConfig
 from lenskit.pipeline.nodes import ComponentInstanceNode
 
 config_dir = Path("pipelines")
+test_dir = Path(__file__).parent
 
 
 def test_load_config():
@@ -99,3 +102,12 @@ def test_apply_config():
     assert isinstance(node, ComponentInstanceNode)
     assert isinstance(node.component, TopNRanker)
     assert node.component.config.n == 100
+
+
+def test_load_override():
+    pipe = Pipeline.load_config(test_dir / "candidate-override.json")
+    assert isinstance(pipe.component("scorer"), SLIMScorer)
+
+    cs = pipe.component("candidate-selector")
+    assert isinstance(cs, TrainingItemsCandidateSelector)
+    assert cs.config.exclude is None
